One last thing
Forgot to add something that might help with the turn after the first to
the third movement. It runs zero yaw twice in one second. Might need to
decrease the time taken up to run it, but it seemed to fix the bug with
robot not turnign on its second turn.